fix(config): improve obsidian daily notes config reader
Before, the `daily-notes.json` file was being parsed into a
`HashMap<String, String>`. However, the deserialization was failing
because on of the fields in that file is `autorun` which is a boolean.
So the deserialization failed and the defaults were being returned
instead of whatever was set in the obsidian config.